Remove `WP_` prefix from grammar rules
In my opinion this prefix isn't needed on the rules. Although in the
beginning I thought that there would be value in the prefix (because we
were mixing rules from different specifications such as HTML) but at
this point it doesn't appear like we'll be expanding the grammar too
much and the prefix isn't helping discriminate block types from other
types.